Split reducers. Dispatch async actions using redux-thunk to handle playback state
