+++ app/scripts/background/actions.js
@@ -9,6 +9,14 @@ export function changePlaybackState() {
   return { type: 'TOGGLE_PLAYBACK' };
 }
 
+export function playStream() {
+  return { type: 'PLAY' };
+}
+
+export function pauseStream() {
+  return { type: 'PAUSE' };
+}
+
 export function resetStream() {
   return { type: 'RESET' };
 }
