+++ app/scripts/background/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import player from './player';
+import track from './track';
+
+export default combineReducers({
+  player,
+  track
+});
