+++ app/scripts/background/reducers/track.js
@@ -1,0 +1,10 @@
+const initialTrackState = {};
+
+export default function track(state = initialTrackState, action) {
+  switch (action.type) {
+    case 'SET_TRACK':
+      return action.data;
+    default:
+      return state;
+  }
+}
