+++ package.json
@@ -31,6 +31,7 @@
     "react-preload": "^0.5.0",
     "react-redux": "^5.0.3",
     "react-tap-event-plugin": "^2.0.1",
-    "redux": "^3.6.0"
+    "redux": "^3.6.0",
+    "redux-thunk": "^2.2.0"
   }
 }
